Move the C++ wrappter to `taichi/cpp`
